4th Commit

+++ Client/src/components/HomePage/Cart.jsx
@@ -1,41 +1,24 @@
 import React, { useState, useEffect } from "react";
 import "../../assets/scss/cart.scss";
+
 import { ViewCart, DeleteCartDetail } from "../../services/UserSevice";
-import { Link } from "react-router-dom";
+import { Link, useNavigate } from "react-router-dom";
+import LoadingAnimation from "../common/LoadingAnimation";
 const Cart = () => {
-  const [items, setItems] = useState([
-    {
-      id: 1,
-      name: 'VINOLI 56" Modern Sofa, Small Corduroy Couch Deep Seat',
-      image: "https://readdy.ai/api/search-image?query=Modern%20beige%20small%20corduroy%20couch%20with%20deep%20seat&width=120&height=120&seq=1&orientation=squarish",
-      originalPrice: 299.0,
-      price: 259.0,
-      savePercent: 20,
-      quantity: 1,
-      size: "S",
-      color: "Beige White",
-      remaining: 2,
-      selected: true,
-    },
-    {
-      id: 2,
-      name: "Fabric Recliner Chair Single Sofa",
-      image: "https://readdy.ai/api/search-image?query=Modern%20teal%20fabric%20recliner%20chair%20single%20sofa&width=120&height=120&seq=2&orientation=squarish",
-      originalPrice: 140.0,
-      price: 109.0,
-      savePercent: 20,
-      quantity: 1,
-      size: "S",
-      color: "White/Chocolate",
-      remaining: 4,
-      selected: true,
-    },
-  ]);
   const [UserInfo, setUserInfo] = useState({});
-  const [couponCode, setCouponCode] = useState("TAKE100");
-  const [appliedCoupon, setAppliedCoupon] = useState(true);
+  const navigate = useNavigate();
   const [CartInfo, setCart] = useState(null);
   const [CartState, setState] = useState(null);
+  const [Loading, setLoading] = useState(true);
+  const handleCheckout = () => {
+    const selectedItems = CartInfo?.cartDetailList.filter((item) => item.selected);
+    navigate("/checkout", { state: { checkoutItems: selectedItems, userInfo: UserInfo } });
+  };
+  const selectedItemsCount = CartInfo?.cartDetailList.filter((item) => item.selected).length;
+  const calSavePrice = (salePrice, price) => {
+    let sale = ((salePrice - price) / price) * 100;
+    return Math.round(sale);
+  };
   useEffect(() => {
     const temp = localStorage.getItem("user");
     const parse = JSON.parse(temp);
@@ -46,11 +29,17 @@ const Cart = () => {
   }, [CartState]);
   const handleCart = async (info) => {
     const res = await ViewCart(info);
-    console.log(res.data.data);
-    setCart(res.data.data);
+    console.log(res);
+    if (res.data.data.cartDetailList != undefined) {
+      const selectItems = res.data.data.cartDetailList.map((item) => ({ ...item, selected: false }));
+      const temp = res.data.data;
+      setCart({ ...temp, cartDetailList: selectItems });
+    }
+    setLoading(false);
   };
+
   const handleDelete = async (id) => {
-    const res = await DeleteCartDetail(id);
+    await DeleteCartDetail(id);
     const data = await ViewCart(UserInfo.id);
     if (data != null) {
       setCart(data.data.data);
@@ -69,7 +58,6 @@ const Cart = () => {
       return item;
     });
     const total = changeQuantity.reduce((sum, item) => sum + item.quantity, 0);
-    console.log(total);
     setCart({ ...CartInfo, sumQuantity: total, cartDetailList: changeQuantity });
   };
   const handleQuantityDown = (id, newQuantity) => {
@@ -81,19 +69,32 @@ const Cart = () => {
       return item;
     });
     const total = changeQuantity.reduce((sum, item) => sum + item.quantity, 0);
-    console.log(total);
     setCart({ ...CartInfo, sumQuantity: total, cartDetailList: changeQuantity });
   };
-  const handleToggleSelect = (id) => setItems(items.map((item) => (item.id === id ? { ...item, selected: !item.selected } : item)));
+  const handleToggleSelect = (id) => {
+    const setSelected = CartInfo?.cartDetailList?.map((item) => (item.cartDetailId === id ? { ...item, selected: !item.selected } : item));
+    setCart({ ...CartInfo, cartDetailList: setSelected });
+  };
 
-  const calculateSubtotal = () => items.filter((item) => item.selected).reduce((total, item) => total + item.price * item.quantity, 0);
-
+  const calculateSubtotal = () =>
+    CartInfo?.cartDetailList
+      ?.filter((item) => item.selected)
+      .reduce((total, item) => {
+        const calPrice = item.productVariant.salePrice != 0 ? item.productVariant.salePrice : item.productVariant.price;
+        return total + calPrice * item.quantity;
+      }, 0);
   const subtotal = calculateSubtotal();
-  const shippingCost = "TBA";
-  const discount = -100.0;
-  const total = subtotal + discount;
-  const selectedItemsCount = items.filter((item) => item.selected).length;
-
+  const total = () =>
+    CartInfo?.cartDetailList
+      ?.filter((item) => item.selected)
+      .reduce((total, item) => {
+        return total + item.productVariant.price * item.quantity;
+      }, 0);
+  const totalPrice = total();
+  const reducePrice = totalPrice - subtotal;
+  if (Loading) {
+    return <LoadingAnimation />;
+  }
   return (
     <div className="cart-container">
       {CartInfo && CartInfo?.cartDetailList?.length != 0 ? (
@@ -107,29 +108,27 @@ const Cart = () => {
                     ← Tiếp tục mua sắm
                   </Link>
                 </div>
-
                 <div className="cart-items">
-                  {CartInfo.cartDetailList?.map((item, index) => (
+                  {CartInfo?.cartDetailList?.map((item, index) => (
                     <div key={index} className="cart-item">
                       <div className="cart-item-left">
-                        <input type="checkbox" checked={item.selected} onChange={() => handleToggleSelect(item.id)} />
+                        <input type="checkbox" checked={item.selected} onChange={() => handleToggleSelect(item.cartDetailId)} />
                         <img src={item.productVariant.productImage[0]} alt={item.name} className="item-img" />
                       </div>
 
                       <div className="cart-item-content">
                         <h3 className="item-name">{item.product.productName}</h3>
                         <div className="item-meta">
-                          <span>Size: {item.size}</span>
+                          <span>Hãng : {item.product.brand}</span>
                           <span>Đơn vị tính: {item.productVariant.calUnit}</span>
                         </div>
                       </div>
-
                       <div className="cart-item-right">
                         <div className="price-area">
                           {item.productVariant.salePrice == 0 ? <span>{item.productVariant.price.toLocaleString("vn-VN", { style: "currency", currency: "VND" })}</span> : <s>{item.productVariant.price.toLocaleString("vn-VN", { style: "currency", currency: "VND" })}</s>}
 
                           <span>{item.productVariant.salePrice != 0 && item.productVariant.salePrice.toLocaleString("vn-VN", { style: "currency", currency: "VND" })}</span>
-                          {item.productVariant.salePrice != 0 ? <div className="save-text">Save {item.savePercent}%</div> : <div></div>}
+                          {item.productVariant.salePrice != 0 ? <div className="save-text">Save {calSavePrice(item.productVariant.price, item.productVariant.salePrice)}%</div> : <div></div>}
                         </div>
 
                         <div className="quantity-area">
@@ -140,11 +139,13 @@ const Cart = () => {
                             <button onClick={() => handleQuantityUp(item.cartDetailId, item.quantity)}>+</button>
                           </div>
                         </div>
-                        {console.log(item.quantity)}
+
                         <button
                           className="remove-btn"
                           onClick={() => {
-                            handleDelete(item.cartDetailId);
+                            if (confirm("Bạn chắc chắn xóa sản phẩm này không ?\nMọi thao tác trước đó với giỏ hàng của bạn sẽ bị hoàn tác")) {
+                              handleDelete(item.cartDetailId);
+                            }
                           }}
                         >
                           Xóa Khỏi Giỏ Hàng
@@ -158,30 +159,29 @@ const Cart = () => {
 
             <div className="cart-right">
               <div className="summary-box">
-                <h3>Coupon</h3>
-                <input type="text" value={couponCode} onChange={(e) => setCouponCode(e.target.value)} placeholder="Enter coupon" />
-                {appliedCoupon && (
-                  <p>
-                    Coupon <strong>{couponCode}</strong> applied - Save $100!
-                  </p>
-                )}
-
                 <div className="summary-info">
-                  <h3>Summary</h3>
-                  <div>Subtotal: ${subtotal.toFixed(2)}</div>
-                  <div>Shipping: {shippingCost}</div>
-                  <div>Discount: ${discount.toFixed(2)}</div>
-                  <div className="final-total">Total: ${total.toFixed(2)}</div>
+                  <h3>Tổng Quan</h3>
                 </div>
               </div>
-
               <div className="cart-summary">
-                <div>{selectedItemsCount} items selected</div>
                 <div className="cart-total">
-                  <span>Total: </span>
-                  <strong>${total.toFixed(2)}</strong>
+                  <span>Tạm tính giỏ hàng : </span>
+                  <strong>{totalPrice.toLocaleString("vn-VN", { style: "currency", currency: "VND" })}</strong>
                 </div>
-                <button className="checkout-btn">Thanh Toán Đơn Hàng</button>
+                <div>{selectedItemsCount} Sản Phẩm Đã Chọn</div>
+                <div> Tổng tiền giảm giá {reducePrice.toLocaleString("vn-VN", { style: "currency", currency: "VND" })}</div>
+                <div className="cart-total">
+                  <span>Tổng tiền : </span>
+                  <strong>{subtotal.toLocaleString("vn-VN", { style: "currency", currency: "VND" })}</strong>
+                </div>
+                <div
+                  className="checkout-btn"
+                  onClick={(e) => {
+                    handleCheckout();
+                  }}
+                >
+                  Thanh Toán Đơn Hàng
+                </div>
               </div>
             </div>
           </div>
